move generate typescript types into cli and as a clie command
