+++ server/cli/src/lib.rs
@@ -11,3 +11,6 @@ pub use helpers::*;
 
 mod plugins;
 pub use plugins::*;
+
+mod generate_typescript_types;
+pub use generate_typescript_types::*;
